requirements/base.txt
#
# Django Core
Django==1.10.5

#
# Connections / Libs / Etc
pytz==2016.7
django-redis==4.6.0
redis>=2.10.5
requests[security]==2.12.4
psycopg2==2.6.2
celery[redis]==3.1.25
-e git+https://github.com/bleib1dj/boto.git@develop#egg=boto
auth0-python==2.0.1
Logentries==0.17
amqp<2.0,>=1.4.5
elasticsearch>=2.0.0,<3.0.0
django_neomodel==0.0.3

#
# Django Contrib
django-environ==0.4.1
django-annoying==0.10.3
django-braces==1.10.0
django-cors-headers==1.3.1
django-filter==1.0.0
django-localflavor==1.3
django-model-utils==2.6
django-clear-cache==0.3
whitenoise==3.2.2
django-organizations==0.8.2
django-revproxy==0.9.13
django-extensions==1.7.4
django-crum==0.7.1
django-material==0.11.0
django-address==0.1.5
django-storages==1.1.8

#
# DRF
djangorestframework==3.5.3
drfdocs==0.0.11
djangorestframework-jwt==1.8.0
django-rest-auth==0.8.2
djangorestframework-jsonapi==2.1.1
rest_framework_auth0==0.2.0
dry-rest-permissions==0.1.8

#
# Everything Else
Unipath==1.1
Markdown==2.6.1
Pillow==3.1.0
WebOb==1.4
argparse==1.2.1
certifi==2016.2.28
six==1.10.0
textblob==0.9.1
urllib3==1.19.1
wsgiref==0.1.2
bcrypt==2.0.0
pyyaml==3.11
unidecode==0.04.19
gwpy==0.1
pycbc-glue
pydub==0.16.5
beautifulsoup4==4.4.1
nltk==3.2.1
oauthlib==1.0.1
stripe==1.32.2
-e git+https://github.com/bleib1dj/django-oauth-toolkit.git@master#egg=oauth2_provider
bleach==1.4.2
us==0.9.0
python-intercom==2.1.1
sumy==0.4.0
imagehash==1.0
pyicu==1.9.2
python-google-places==1.1.0
html5lib==1.0b8
python-amazon-simple-product-api==2.1.0
shortuuid==0.4.2
scipy==0.17.1
awscli==1.10.34
numpy==1.11.0
